Extract theme toggle component
